Add private dependencies
These are both hosted on GitHub Packages. They're authenticated using a GitHub personal token in ~/.npmrc locally, and with the `GPR_TOKEN` org secret in CI.

npm knows to pull them from there by the declaration in .npmrc.

CI needs to be modified so that it respects the location of scoped packages, but in this case we also added dependency caching